refactor: MainPage Channel Container 코드 스플리팅 및 Channel 컴포넌트 네이밍 수정
