+++ src/pages/MainPage/_compoenets/Content/ChannelContainer/Channel/Channel.type.tsx
@@ -1,5 +1,5 @@
 import { Channel } from '@/types/channel';
 
-export interface ChannelItemProps {
+export interface ChannelProps {
   channel: Channel;
 }
