+++ src/pages/MainPage/_compoenets/Content/ChannelContainer/Channel/TagContainer/TagContainer.type.tsx
