+++ src/pages/MainPage/_compoenets/Content/ChannelContainer/ChannelContainer.hook.tsx
@@ -2,13 +2,11 @@ import useGetAllChannels from '@/apis/hooks/useGetAllChannels';
 import useMusicStore from '@/store/useMusicStore';
 
 export default function useChannelContainer() {
-  const { data: channels, isLoading, isError } = useGetAllChannels();
+  const { data: channels } = useGetAllChannels();
   const { music: currentMusic } = useMusicStore();
 
   return {
     channels,
-    isLoading,
-    isError,
     currentMusic,
   };
 }
